chore: 금융상품, 회원가입, 로그인, 버킷리스트 랜딩페이지 메타데이터 작성
우선 title과 description만 작성했습니다.
