+++ app/auth/signup/layout.tsx
@@ -1,4 +1,10 @@
+import { Metadata } from 'next';
 import SignUpFormProvider from './_components/SignupFormProvider';
+
+export const metadata: Metadata = {
+  title: 'PORKO | 회원가입',
+  description: 'PORKO 회원가입 페이지'
+};
 
 const AuthLayout = ({ children }: { children: React.ReactNode }) => {
   return (
