+++ app/financial-product/comparison/_components/ComparisonSection.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import Text from '@/components/ui/Text';
 import { useEffect } from 'react';
 import { useQueryString } from '@/shared/hooks/useQueryString';
@@ -6,6 +8,7 @@ import { getCardsToCompare } from '@/service/api/financial-product/cards';
 import Spinner from '@/components/Spinner';
 import FlexBox from '@/components/ui/FlexBox';
 import CardsToCompare from './CardsToCompare';
+import BottomButton from './BottomButton';
 
 const QUERY_KEY = 'card';
 
@@ -14,9 +17,9 @@ const ComparisonSection = () => {
     queryKey: ['cardsToCompare'],
     queryFn: getCardsToCompare
   });
-  const { searchParams, pathname, router, params } = useQueryString();
+  const { pathname, router, params, queryValues } = useQueryString();
 
-  const selectedCards = searchParams.getAll(QUERY_KEY);
+  const selectedCards = queryValues(QUERY_KEY);
 
   useEffect(() => {
     if (selectedCards.length > 2) {
@@ -38,31 +41,36 @@ const ComparisonSection = () => {
   };
 
   return (
-    <section className='px-20'>
-      <Text variant='h2' sizes='20' weight='700' className='mb-4'>
-        비교할 카드 선택하기
-      </Text>
-      <Text variant='p'>
-        최대 <Text weight='700'>2개</Text>까지만 선택할 수 있어요
-      </Text>
-      {isPending ? (
-        <FlexBox justifyContent='center' className='mt-20'>
-          <Spinner />
-        </FlexBox>
-      ) : (
-        <ul className='mt-20 flex flex-col gap-[1.2rem]'>
-          {cardsToCompare?.map((card) => {
-            const isSelected = selectedCards.some((c) => c === card.id);
+    <>
+      <section className='px-20'>
+        <Text variant='h2' sizes='20' weight='700' className='mb-4'>
+          비교할 카드 선택하기
+        </Text>
+        <Text variant='p'>
+          최대 <Text weight='700'>2개</Text>까지만 선택할 수 있어요
+        </Text>
+        {isPending ? (
+          <FlexBox justifyContent='center' className='mt-20'>
+            <Spinner />
+          </FlexBox>
+        ) : (
+          <ul className='mt-20 flex flex-col gap-[1.2rem]'>
+            {cardsToCompare?.map((card) => {
+              const isSelected = selectedCards.some((c) => c === card.id);
 
-            return (
-              <li key={card.id}>
-                <CardsToCompare isSelected={isSelected} onSelect={onSelect} card={card} />
-              </li>
-            );
-          })}
-        </ul>
+              return (
+                <li key={card.id}>
+                  <CardsToCompare isSelected={isSelected} onSelect={onSelect} card={card} />
+                </li>
+              );
+            })}
+          </ul>
+        )}
+      </section>
+      {selectedCards.length >= 2 && (
+        <BottomButton path='/financial-product/comparison/select-category'>비교하기</BottomButton>
       )}
-    </section>
+    </>
   );
 };
