+++ app/financial-product/comparison/page.tsx
@@ -1,24 +1,20 @@
-'use client';
-
 import { IsBackHeader } from '@/components/header';
 import FilteringSection from './_components/FilteringSection';
 import ComparisonSection from './_components/ComparisonSection';
-import BottomButton from './_components/BottomButton';
-import { useQueryString } from '@/shared/hooks/useQueryString';
+import { Metadata } from 'next';
+
+export const metadata: Metadata = {
+  title: 'PORKO | 상품 비교',
+  description: 'PORKO 상품 비교 페이지'
+};
 
 const ComparisonPage = () => {
-  const { queryValues } = useQueryString();
-  const selectedCards = queryValues('card');
-
   return (
     <>
       <IsBackHeader title='상품비교' href='./?tab=신용카드' defaultColor='#fff' />
       <main className='bg-gray-50 pb-[13.2rem]'>
         <FilteringSection />
         <ComparisonSection />
-        {selectedCards.length >= 2 && (
-          <BottomButton path='/financial-product/comparison/select-category'>비교하기</BottomButton>
-        )}
       </main>
     </>
   );
